More experiments with the tokenizer
