+++ tests/temp.spec.js
@@ -13,12 +13,16 @@ import { buildSchemas } from '../validator/schema/init'
 import ColumnSplicer from '../parser/columnSplicer'
 import ParsedHedGroup from '../parser/parsedHedGroup'
 import { HedStringTokenizer } from '../parser/tokenizer'
-import { HedStringTokenizerNew } from '../parser/tokenizerNew'
+import { HedStringTokenizerOld } from '../parser/tokenizerOld'
 
 describe('HED string parsing', () => {
   it('should include each group as its own single element', () => {
-    const hedString = '(xy)'
-    const tok = new HedStringTokenizerNew(hedString)
+    //const hedString = "Action/Move/Flex,(Relation/Spatial-relation/Left-side-of,Action/Move/Bend,Upper-extremity/Elbow),Position/X-position/70 px,Position/Y-position/23 px"
+    //const hedString = 'x/y w/z'
+    const hedString = '(r,z)'
+    //const hedString = 'r,'
+    //const hedString = 'r,y'
+    const tok = new HedStringTokenizer(hedString)
     const [tagSpecs, groupBounds, tokenizingIssues] = tok.tokenize()
     assert.isEmpty(Object.values(tokenizingIssues).flat(), 'Parsing issues occurred')
     //const hedString = 'Action/Move/My-flex,(Relation/Spatial-relation/Left-side-of,Action/Move/My-bend,Upper-extremity/My-elbow),Position/X-position/70 m,Position/Y-position/23 m'
